refactor(rpc): :art: changes based on PR feedback
Move some classes to their own files, place types in the correct directories etc based on the PR feedback from Yash
