+++ src/server/controllers/relay/rpc/RPCResponse.ts
@@ -1,0 +1,6 @@
+export abstract class RPCResponse {
+  constructor(
+    public id = 1,
+    public jsonprc = "2.0",
+  ) {}
+}
