+++ src/server/controllers/relay/rpc/index.ts
@@ -1,0 +1,2 @@
+export * from "./RPCBadRequest";
+export * from "./RPCInternalServerError";
